+++ src/Fieldtypes/Fieldtype.php
@@ -36,7 +36,7 @@ abstract class Fieldtype
     /**
      * A list of supported parameters for the fieldtype
      */
-    public function parameters(Field $field = null): array
+    public function parametersForField(Field $field = null): array
     {
         return [];
     }