+++ src/Fieldtypes/Grid.php
@@ -9,130 +9,26 @@ use Expressionengine\Coilpack\Contracts\GeneratesGraphType;
 use Expressionengine\Coilpack\Contracts\ListsGraphType;
 use Expressionengine\Coilpack\FieldtypeManager;
 use Expressionengine\Coilpack\FieldtypeOutput;
+use Expressionengine\Coilpack\Fieldtypes\Presenters\GridPresenter;
 use Expressionengine\Coilpack\Models\Channel\ChannelField;
 use Expressionengine\Coilpack\Models\FieldContent;
 use Expressionengine\Coilpack\Support\Parameter;
-use Expressionengine\Coilpack\View\FilteredParameterValue;
-use Illuminate\Support\Facades\DB;
 
 class Grid extends Fieldtype implements GeneratesGraphType, ListsGraphType
 {
+    protected $presenter;
+
+    public function __construct(string $name, $id = null)
+    {
+        parent::__construct($name, $id);
+        $this->presenter = new GridPresenter;
+    }
+
     public function apply(FieldContent $content, array $parameters = [])
     {
-        $data = $this->loadData($content, $parameters);
+        $data = $this->presenter->present($content, $parameters);
 
         return FieldtypeOutput::for($this)->value($data);
-    }
-
-    protected function loadData(FieldContent $content, array $parameters = [])
-    {
-        if (! $content->hasAttribute('entry_id')) {
-            return [];
-        }
-
-        $isFluid = $content->hasAttribute('fluid_field');
-
-        $parameters = array_merge([
-            'orderby' => 'row_order',
-            'sort' => 'asc',
-            'limit' => 100,
-            'offset' => 0,
-        ], $parameters);
-
-        $data = $content->entry->isPreview() ? $this->loadFromPreview($content, $parameters) : $this->loadFromDatabase($content, $parameters);
-
-        $columns = $content->field->gridColumns;
-
-        // maybe we want to persist $data from the query before we do any filtering?
-        $data = $data->filter(function ($row) use ($isFluid, $content) {
-            return ! $isFluid || $row->fluid_field_data_id == $content->fluid_field_data_id;
-        })->map(function ($row) use ($columns, $content) {
-            $columns->each(function ($column) use ($row, $content) {
-                $row->{$column->col_name} = new FieldContent(
-                    array_merge($content->getAttributes(), [
-                        'data' => $row->{'col_id_'.$column->col_id},
-                        'grid_row_id' => $row->row_id,
-                        'grid_col_id' => $column->col_id,
-                        'fieldtype' => app(FieldtypeManager::class)->make($column->col_type),
-                    ])
-                );
-            });
-
-            return $row;
-        });
-
-        return $data;
-    }
-
-    protected function loadFromPreview($content, array $parameters = [])
-    {
-        $isFluid = $content->hasAttribute('fluid_field');
-        $fluidFieldId = ($isFluid) ? $content->fluid_field_data_id : 0;
-
-        $rows = collect($content->data['rows'] ?? [])->map(function ($value, $key) use ($fluidFieldId) {
-            $id = str_replace('row_id_', '', $key);
-
-            return (object) array_merge([
-                'row_id' => $id,
-                'fluid_field_data_id' => $fluidFieldId,
-            ], $value);
-        });
-
-        if ($parameters['fixed_order'] ?? false) {
-            // fixed_order
-        } else {
-            $rows = ($parameters['sort'] == 'asc') ? $rows->sortBy($parameters['orderby']) : $rows->sortByDesc($parameters['orderby']);
-        }
-
-        // Filter by row ids
-        if ($parameters['row_id'] ?? false) {
-            $rows = $rows->where('row_id', '=', $parameters['row_id']);
-        }
-
-        // Handle offset
-        if ($parameters['offset']) {
-            $rows = $rows->skip($parameters['offset']);
-        }
-
-        $rows = $rows->take($parameters['limit']);
-
-        return $rows;
-    }
-
-    protected function loadFromDatabase(FieldContent $content, array $parameters = [])
-    {
-        $isFluid = $content->hasAttribute('fluid_field');
-        $fluidFieldId = ($isFluid) ? $content->fluid_field_data_id : 0;
-
-        $tableName = "channel_grid_field_{$content->field->field_id}";
-        $query = DB::connection('coilpack')
-            ->table($tableName)
-            ->where('entry_id', $content->entry_id)
-            ->where('fluid_field_data_id', '=', $fluidFieldId); // $fluidFieldId);
-
-        // Handle ordering
-        if ($parameters['fixed_order'] ?? false) {
-            $query->whereIn('row_id', $parameters['fixed_order']);
-            $query->orderByRaw('FIELD(row_id, '.implode(',', $parameters['fixed_order']).')');
-        } else {
-            $direction = ($parameters['sort'] == 'asc') ? 'asc' : 'desc';
-            $query->orderBy($parameters['orderby'], $direction);
-        }
-
-        // Filter by row ids
-        if ($parameters['row_id'] ?? false) {
-            $rowIds = new FilteredParameterValue($parameters['row_id']);
-            $rowIds->filterQueryWithColumn($query, 'row_id');
-        }
-
-        // Handle offset
-        if ($parameters['offset']) {
-            $query->skip($parameters['offset']);
-        }
-
-        $data = $query->take($parameters['limit'])->get();
-
-        return $data;
     }
 
     public function generateGraphType(ChannelField $field)
@@ -159,56 +55,26 @@ class Grid extends Fieldtype implements GeneratesGraphType, ListsGraphType
         ]);
     }
 
-    public function parameters(Field $field = null): array
+    public function parametersForField(Field $field = null): array
     {
-        return [
-            new Parameter([
-                'name' => 'fixed_order',
-                'type' => 'string',
-                'description' => 'Order rows in a fixed order of row IDs',
-            ]),
-            new Parameter([
-                'name' => 'limit',
-                'type' => 'integer',
-                'description' => 'Limits the number of rows',
-            ]),
-            new Parameter([
-                'name' => 'offset',
-                'type' => 'integer',
-                'description' => 'Offsets the number of rows',
-            ]),
-            new Parameter([
-                'name' => 'orderby',
-                'type' => 'string',
-                'description' => 'Order rows by a specific column',
-            ]),
-            new Parameter([
-                'name' => 'row_id',
-                'type' => 'string',
-                'description' => 'Only rows for the given IDs. Multiple rows may be specified by separating them with a pipe character',
-            ]),
-            new Parameter([
-                'name' => 'search',
-                'prefix' => $field->field_name,
-                'description' => 'Search for rows matching a certain criteria',
-                'type' => function () use ($field) {
-                    return $field->gridColumns->flatmap(function ($column) {
-                        return [
-                            new Parameter([
-                                'name' => $column->col_name,
-                                'type' => 'string',
-                                'description' => $column->col_instructions,
-                            ]),
-                        ];
-                    })->toArray();
-                },
-            ]),
-            new Parameter([
-                'name' => 'sort',
-                'type' => 'string',
-                'description' => 'Specifies the direction of the sorting',
-                'defaultValue' => 'asc',
-            ]),
-        ];
+        $parameters = $this->presenter->defineParameters();
+        $parameters[] = new Parameter([
+            'name' => 'search',
+            'prefix' => $field->field_name,
+            'description' => 'Search for rows matching a certain criteria',
+            'type' => function () use ($field) {
+                return $field->gridColumns->flatmap(function ($column) {
+                    return [
+                        new Parameter([
+                            'name' => $column->col_name,
+                            'type' => 'string',
+                            'description' => $column->col_instructions,
+                        ]),
+                    ];
+                })->toArray();
+            },
+        ]);
+
+        return $parameters;
     }
 }