+++ src/Fieldtypes/Presenters/Presenter.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Expressionengine\Coilpack\Fieldtypes\Presenters;
+
+use Expressionengine\Coilpack\Models\FieldContent;
+
+abstract class Presenter
+{
+    abstract public function present(FieldContent $content, $arguments);
+}