+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Page variables to Channel Entries Tag results
 - Global Variables for logged_in and logged_out
 - Better error handling for incomplete installations
+- Support for Grid parameters on LivePreview data
 
 ### Fixed
 
@@ -56,6 +57,7 @@
 - Cleanup and simplify Template library for rendering Twig and Blade templates
 - No longer overriding ExpressionEngine's `base_url` with Laravel's `APP_URL` unless `base_url` is not set.
 - Channel Entries Tag now has a default ordering to match native Tag.  Sticky entries are first followed by those with the latest entry_date.
+- Fieldtype `parameters()` has been renamed to `parametersForField()`
 
 ## [0.1.0] - 2023-02-13